requirements/dev.txt
asgiref==3.10.0
django==5.2.7
djangorestframework==3.16.1
faker==37.12.0
pip==25.1.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
